Initial commit of GEOS-Chem Unit Tester
The GEOS-Chem Unit Tester is a combination of Perl scripts and GNU
Makefiles that repeatedly compile & run GEOS-Chem with various
debugging options.  The user edits an input file with the various
debugging options as well as the list of simulations that he/she
would like to perform.

At present each Unit Test involves of running GEOS-Chem with the
same compilation options for both single processor and multi processor.
We then difference the results.  The user may select debug options,
as well as specify the directory paths in an input file.

We may need to further polish the user interface.  Bob Y. tested this
with the geos5 4x5 RnPbBe simulation and it worked properly.

We shall also add extra GEOS-Chem run directories in the future.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>