Bug fix: List drydep tracers explicitly in ND44 diagnostic
We now need to explicitly list all drydep tracers for the
ND44 diagnostic, or else we will get a difference in the bpch
files in the DRYD-FLX and DRYD-VEL categories.  This bug is
still under investigation, but this is the cheap fix.

Also save new version of UnitTest.input file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>