Add a geosfp_4x5_UCX unit test simulation
Added the run directory runs/geosfp_4x5_UCX to support UCX
unit tests at 4x5 with the GEOS-FP met.  From this directory,
we symlink to the input files in ../chem_inputs/UCX and to
the restart files in ../geos5_4x5_UCX.  The GEOS-FP and
GEOS-5 restart files use the same grid so we can interchange
them.

Also in the runs/geos5_4x5_UCX directory, we now symlink to
the chem mechanism files in ../chem_inputs/UCX.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>