Add gcUtMatrix script to create a unit test matrix web page
The perl/gcUtMatrix script now can read a results.log file
and create a web page showing a table of the results of each
GEOS-Chem unit test.  The template ut_template.html contains
the HTML script, and we replace tokens for the various colors
indicating the status of each unit test.

NOTE: We still need to replace the version and date in the
HTML file, but we'll work on that later.

ALSO NOTE: We have added placeholders in the runs directory
for unit tests that do not yet exist.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>