Add line for PSC restart file to input.geos
We now allow users to specify the location of the PSC restart file in the
Aerosol Menu of input.geos. The input.geos files in all unit tester run
directories have been updated.

A few cosmetic updates have also been made
to the input.geos files so that they are more consistent with one another,
thus facilitating ediffing in the future.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>