Bug fix in globchem.dat files
In a previous commit, we had clobbered the globchem.dat files
pertaining to the fullchem, soa, and soa_svpoa simulations.
We have now restored those, and have made the fix for the
RIO2 reaction (i.e. removing the extra 0.5GLYX product).

Also, we now make sure that the ProTeX headers in the
globchem.dat files say "GEOS-Chem Chemical Transport Model.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>