Edit HEMCO_Config files to read local soil NOx restart files
The HEMCO_Config files now point to soil NOx restart files
in the various run directories.  (Where possible, these are
symbolically linked to each other in order to save disk
space.).

We also updated the Makefiles in the various run directories
to account for the new HEMCO diagnostic files.

We have converted the bpch soil NOx restart files to
netCDF for HEMCO input.  We have removed the bpch files.

Also updated the .gitignore files to ignore the HEMCO
restart files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>