Add "extrafileclean" target to run directory Makefiles
We now have added the "extrafileclean: target to the run dir
Makefiles.  This target will remove the input.geos and
HEMCO_Config.rc files that are created for each unit test.

The "extrafileclean" target is only called from perl/cleanFiles,
which is used to completely clean up all unit test files.

The "fileclean" target is called when you make "superclean".
This is also called when starting a new unit test.  We cannot
remove input.geos and HEMCO_Config.rc in "fileclean" because
that would remove the files before they ever get used.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>