Now read HEMCO input file named "HEMCO_Config.rc"
In various run directories, make sure the name of the HEMCO
Config file is now "HEMCO_Config.rc".  This file is created
from the HEMCO_Config.template file (or symbolic link to a
HEMCO_Config.template file) in each run directory.

Also updated the .gitignore files in the various run directories
to ignore HEMCO outputs, as well as the HEMCO_Config.rc files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>