Now allow ending time to be either YYYYMMDDhh or YYYYMMDDhhmm
We have modified the gcUnttTest script to allow the ending
time to be in either YYYYMMDDhh format (10-digits) or
YYYYMMDDhhmm format (12 digits).

For some unit tests that take a long time to run (e.g.
geos5_2x25_fullchem), running for 1 chemistry timestep (30 mins)
will make the unit test process go faster.

Also improved comments in the gcUnitTest script.
Modified the UnitTest input files accordingly.

Because we only run for 1 chemistry timestep, we have hardwired
the HEMCO restart file name so that it uses the starting time.
In reality, the HEMCO restart file is timestamped with the
YYYYMMDDhhmm that is 1 chemistry timestep prior to the end of
the simulation.  We need to think about how to make this more
general later on.  The solution will probably involve a call
to the Julian day routine to compute the proper date & time.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>