Add FINN file info to HEMCO_Config.template
We have now added the FINN data file info in the extension
section of runs/chem_inputs/fullchem/HEMCO_Config.template.
This file is also symlinked to various other directories.

NOTE: FINN is turned off for now, as it has not yet been
validated with a benchmark simulation in the standard G-C.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>