Add {MET} token to the HEMCO_Config.template.tagO3 file
In order to be able to select either the GEOS-4 (30L) or the
GEOS-5/MERRA/GEOS-FP (47L) file with the O3 prod and loss data,
we had to add the {MET} token to the filename in the
runs/chem_inputs/specialty/HEMCO_Config.template.tagO3 file.

Also add a couple of symbolic links in the HEMCO data directory
so that geosfp and merra will link to the O3 prod and loss file
named with geos5.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>