Update HEMCO_Config.template; use new soilNOx files.
Christoph Keller added some fixes to HEMCO, in both the configuration
files and in the data files.  Updated the HEMCO_Config.template
files for the full-chemistry, SOA, and aerosol simulations.

We also now point to a new soil restart file which has units in kg/m3
instead of in ngN/m2.  That is a more standard unit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>