Add mass tuning factor HEMCO config files for fullchem, soa
GCAP and GEOS-4 unit tests for fullchem and SOA won't run unless
we specify the Mass Tuning Factor in the HEMCO extension file.

NOTE: For unit testing it is fine to leave the mass tuning factor
in the HEMCO_Config.rc file for GEOS-5, MERRA, or GEOS-FP.  But if
you copy the HEMCO_Config.rc file for use elsewhere you should delete
that line.  It will be determined automatically for those met fields.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>