Use separate HEMCO_Config.template files for GEOS-4 and GCAP fullchem
The GEOS-4 and GCAP simulations require the "mass tuning factor"
to be specified in the HEMCO_Config.template file.  However, other
simulations (GEOS-5, MERRA, GEOS_FP) do not normally use this.

Therefore, we now link to separate versions of the HEMCO config
files in the GEOS-4 and GCAP fullchem and aerosol-only run dirs.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>