Add perl/gcCopyRunDirs script to generate GC run directories
Created a new script called gcCopyRunDirs.  This script is based off
gcUnitTest.  It reads a UnitTest.input file and then copies the
selected run directories to a root directory path that is specified.
The input.geos and HEMCO_Config.rc files in each run directory are
created from template files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>