Now change 2005 to 2006 so that we can use NEI2011 emissions
We now have scale factors to scale the NEI2011 emissions from the
base year of 2011 down to 2006.  Now change the unit test dates
from 2005 to 2006 for GEOS-5, GCAP, and MERRA met fields.

This will make sure that the proper settings will be activated when
GEOS-Chem users copy the run directories with gcCopyRunDirs.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>