More minor updates in the perl/Makefile
Added the following changes to perl/Makefile:

(1) Now send the HEMCO.log file to the run directory instead of to
    the log directory.  If you are using the perl/Makefile in a copied
    run dir, then the log & run dirs are the same.  But when running
    a unit test, this causes the log files to be sent to the wrong
    place.

(2) Minor bug fixes in the ifndef statements.  Make sure to add the
    variable name after ifndef.

(3) Cosmetic changes.  Updated comments and made separators to make
    it easier to distinguishe between sections of the Makefile.

(4) Added a help screen.  The user can type "make help" for more
    information about all the Makefile targets

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>