Now use YYYYMMDDhhmm format in trac_avg filenames
We will now use trac_avg.[MET]_[RES]_[SIM].YYYYMMDDhhmm for the binary
punch file name, instead of .YYYYMMDDhh. We include "mm" in the filename
for the trac_rst files and HEMCO uses the YYYYMMDDhhmm format to label the diagnostic and
restart files. The timestamp format is now consistent between all output
files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>