Restore MIX as the default Asian emission inventory over Streets
Qiang Zhang and Meng Li provided plots that show the difference in SO2
emissions that we observed in the v10-01h benchmarks is consistent with
their comparisons of Streets and MIX at 4x5. Therefore, there doesn't
appear to be a bug in the MIX emissions, so it has been restored as the
default Asian emission inventory.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>