Now use tokens to set the HEMCO verbose and warnings settings.
You now have the option of setting the HEMCO verbose and warnings
levels via tokens from the UnitTest.* input and CopyRunDir.* input
files.

For GEOS-Chem unit tests, if VERBOSE and WARNINGS are not specified,
default settings of 3 and 3 will be used.  This provides maximum debug
output (which is what you want for unit tests).

For generating GEOS-Chem run directories, if VERBOSE and WARNINGS are
not specified, then default settings of 0 and 1 will be used.  This
will minimize debug output for production runs.

We have modified the gcUnitTest, gcCopyRunDirs scripts, as well as the
various HEMCO_Config.template* files accordingly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>