Rename HEMCO restart files so that they contain the date
Changed the HEMCO restart file names from initial_hemco_restart*.nc
to HEMCO_restart.YYYYMMDDhhmm.nc, in order to match the change that
was made to GEOS-Chem v10-01 provisional release.

Also updated the .gitignore files not to ignore files beginning
with HEMCO_restart*.

Need to modify HEMCO configuration files but this will be done
in the subsequent commit.