Add VERBOSE and WARNINGS to the UnitTest.input files
We now explicitly add the VERBOSE and WARNINGS options to the
UnitTest.* input files.  We now explicitly set these to the max
values of VERBOSE=3 and WARNINGS=3.

NOTE: This is mostly for the sake of completeness.  The gcUnitTest
script will automatically default to VERBOSE=3 and WARNINGS=3 if
these are not specified in the UnitTest.* input files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>