The perl/Makefile now passes CHEM to the build sequence
We now set the values of UCX and CHEM properly for the various
run directories.  The combinations are:

 CHEM=benchmark
 CHEM=SOA
 CHEM=UCX
 CHEM=NOx_Ox_HC_Br

CHEM is currently determined from the simulation type, which is part
of the run directory name.  CHEM is also explicitly passed to the
build sequence.

Also, add a ./ before the executable names in the sp, mp, sprun, and
mprun targets.  Some user environments must specify the ./ explicitly
in order to find a file in the current directory.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>