Remove PARANOX_SUNCOS entries from the HEMCO_Config.template files
Christoph Keller modified the HEMCO PARANOX extension so that it now
computes the SUNCOS and SUNCOS5 variables within the extension.
We no longer need to read the PARANOX_SUNCOS* entries from the HEMCO
restart file.

We have removed the PARANOX_SUNCOS entries from all of the relevant
HEMCO_Config.template files, as well as from the HEMCO restart
files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>