Bug fix in perl/Makefile: set NEST=n when SIM=soa_svpoa
The simulation name soa_svpoa has an underscore in it.  This causes
the script getRunInfo (which is called by perl/Makefile) to think
that svpoa is the nested-grid name instead of part of the simulation
name.

For the time being, the quick fix is to set NEST=n when SIM=soa_svpoa.
Need to think of a more robust solution later.  Perhaps we need
to rename the soa_svpoa to soa-svpoa instead, to avoid confusion with
the underscore.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>