Add logical switch for the HEMCO restart file
We now include a switch HEMCO_RESTART that allows users to turn on/off
using the HEMCO restart file to initialize their simulation. When the HEMCO
restart file is turned off, HEMCO will use default values specified within
the HEMCO extensions. We recommend users turn off using the HEMCO restart
file when they don't have a file that corresponds to the start date of
their simulation. It's better to use the default values for those fields
than to use values for the wrong date.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>