Modify input.geos.template files to save monthly output by default
The Output Menu of all input.geos.template files has been modified so that
we now save monthly output for each month instead of daily output. This
means that we now have a 3 placed on the first day of each month instead of
each day of each month. This should not affect the unit tests because they
are set up to run for <1 day.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>