Now append the SLURM #SBATCH tags to the job script file
We have added several (optional) parameters to the UnitTest.input
file(s) that specify parameters for the SLURM scheduler.  If these
parameters are present, then the proper #SBATCH tags will be written
at the top of the job script file.

Also made a minor modification in the &parse routine (in UtUtils.pm).
&parse now works if there are 2 or more ":" characters in the line.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>