Now create a simple text file to summarize unit test results
On some computer systems, when you are running unit tests in a
computational queue, you cannot push results to a web server.
Therefore, we now also create a simple text file with unit test
results that can be scanned easily.  The file name is specified
by the TXT_GRID tag in the UnitTest.* input files.

We have modified the perl/gcUtWeb, perl/gcUnitTest, and
perl/UtUtils.pm files accordingly to allow the summary of unit
test results to be sent to a text file.

If you type "NONE" (or "None" or "none", etc) in the WEB_GRID tag,
then the summary of unit test results will not be sent to an HTML
file.

Similarly, if you type "NONE" (or "None" or "none", etc) in the
TXT_GRID tag, then the summary of unit test results will not be
sent to the text file.

Also add optional commands for the SLURM scheduler to the various
perl/inputs/UnitTest.* files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>