Add option for acid uptake on dust to input.geos and HEMCO_Config.rc files
In the Aerosol Menu of input.geos, we have added a line for the acid
uptake on dust option developed by Duncan Fairlie. This option is turned
off by default.

An option for DustAlk (dust alkalinity source) has been added to the
HEMCO_Config.rc files. This option applies to the DSTAL1, DSTAL2, DSTAL3,
and DSTAL4 species is turned off by default.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>