Update lightning NOx entries in HEMCO config to use single file for GEOS-5
We had previously split the GEOS-5 files into two files: (1) a geos5.1.0
file for dates prior to Aug 1, 2008 and (2) a geos5.2.0 file for dates
Aug 1,2008 and beyond. This can cause problems when a run spans multiple
months in 2008. To resolve this, we have created a single file for GEOS-5.
The LIGHTNOX entries in HEMCO_Config.template files have been updated
accordingly.

This update also means that we can now use the $MET token in the LIGHTNOX
filename. As such, LNOX is no longer needed and references to this token
have been removed.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>