Add UNIT_TEST_ROOT path to the *.input files
Prior versions of the unit tester assumed that the unit tester would
be placed in a user's home folder.  Several unit test sub-folders
were specified in the *.input files relative to the home folder, e.g.:

   RUN_ROOT         : {HOME}/UT/runs
   JOB_DIR          : {HOME}/UT/jobs
   ... etc ...

On many systems it is necessary to install the unit test directory
on a scratch disk (which usually has faster I/O and much more memory)
instead of in the home folder.  We have therefore modified the various
*.input files and perl scripts to allow a user to specify the unit
tester's root directory independent of the home directory, as
shown below:

   UNIT_TEST_ROOT   : /fast/scratch/disk/UT
   RUN_ROOT         : {UTROOT}/runs
   JOB_DIR          : {UTROOT}/jobs
   ... etc ...

The {UTROOT} token in directory paths will be replaced with the
value specified as "UNIT_TEST_ROOT".

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>