Apply fixes to HEMCO configuration file for soa_svpoa simulations
To make sure proper scale factor is applied to the SVOC emissions in GFED,
we need to add the full tracer name (POG1) to the Scaling_xxx label.

Here we also change all SVOC emissions so that they are assigned to the
POG1 species to better reflect that these emissions are added to the
gas-phase species. Before, they were added to both POA1 (biomass+biofuel)
and POG1 (anthropogenic), but the partitioning by emission type was only
done in carbon_mod.F for debugging purposes and is no longer needed. The
total SVOC emissions from HEMCO are added to the POAEMISS array in
EMISSCARBON (in GeosCore/carbon_mod.F) and the assignment of the SVOC
emissions to the two gas-phase species occurs in routine CHEM_NVOC (also
in GeosCore/carbon_mod.F).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>