Update input.geos.template for RnPbBe to include passive tracer (PASV)
In the Rn-Pb-Be benchmark simulations, we now include a passive tracer
(PASV) to test mass conservation in GEOS-Chem. The input.geos.template
files for the Rn-Pb-Be unit tests have now been updated to include this
additional tracer. We also now point to a different restart file (named
with RnPbBePasv, instead of RnPbBe) that includes the PASV tracer.

The diagnostic settings in the input.geos.template files have also been
updated so that they are more consistent with the settings in the
1-year Rn-Pb-Be benchmark simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>