Update end dates in UnitTest.input files to reflect new timesteps
The default timesteps specified in the input.geos.template files are now
10 minutes for transport and convection and 20 minutes for emissions and
chemistry. The unit tests are set up so that we complete at least one
chemistry timestep. Therefore, we update the UnitTest.input files so that
we now run all unit tests for 20 minutes.

NOTE: The exception to the 20 minute unit tests are 0.25x0.3125 and RRTMG
simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>