Update v11-01.1mon input file in benchmark directories to include Cloud_Diffs option
We now include an option to plot differences of cloud fraction and cloud
optical depth in the 1-month benchmark plotting routines. This update
required the addition of a line for Cloud_Diffs in the v11-01.1mon input
file that we pass to GAMAP routine benchmark_1mon.pro.

Also added the following minor updates to the scripts in the benchmark
directories:
  (1) Make sure we copy all *.fullchem files to the ftp site in copy_1mon
  (2) Pass /No_CloudDiff keyword when we want to generate select benchmark
      plots in plot_1mon.pro. (These sections are currently commented out.)

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>