Turn on ND31 diagnostic for all unit test simulations
GEOS-Chem user Jinhyok Heo reported that the ND31 diagnostic had not
been turned on in the input.geos.template file for the SOA simulation.
This caused the metadata for ND31 not to be written to the diaginfo.dat
and tracerinfo.dat files.  This in turn caused an error when Jinhyok
tried to regrid the restart file (because GAMAP could not interpret
the ND31 surface pressure data).

We have now activated the ND31 surface pressure diagnostic for all
unit test simulations.  This should write out the metadata for GAMAP
properly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>