Remove LKPP option from input.geos.template files
The LKPP option in the Chemistry Menu of input.geos is no longer needed.
With the introduction of FlexChem in v11-01g, SMVGEAR has been removed
and KPP is the only chemical solver available. As such, the LKPP option
is now obsolete.

Also turned off the prod/loss diagnostic (ND65) for the TOMAS simulations.
This diagnostic is currently unavailable with FlexChem, but the GCST
plans on restoring it completing the conversion of GEOS-Chem diagnostics
from BPCH to netCDF format.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>