Update DiffTest Makefile to print out git information
The printruninforef and printruninfodev options now include information
from git, including the branch, last commit message, and commit date.
These print statements are added to the printruninfo* options and not the
printbuildinfo section, where they may be better suited, because the
information will differ for the Ref and Dev runs.

Also updated Makefiles in the difftest/Ref and difftest/Dev directories
so that they are consistent with one another.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>