Consolidate run directories for GEOS-5, GEOS-FP, MERRA, and MERRA-2
The run directories for GEOS-5, GEOS-FP, MERRA, and MERRA-2 were
essentially identical because those met fields have the same vertical
and horizontal resolutions. The only differences in the run directories
was the use of the met field name in the input.geos.template files.
The GEOS-5, GEOS-FP, MERRA, and MERRA-2 4x5 and 2x2.5 run directories
have now been consolidated into 4x5_{SIM} and 2x25_{SIM} directories.
The input.geos.template files now include a {MET} token, which is replaced
by the met field via gcUnitTest, gcCopyRunDirs, or gcCreateDiff perl
scripts. The desired met field is specified in the UnitTest.input,
CopyRunDirs.input, and DiffTest.input file. When a user creates a run
directory, the met field will be inserted into the directory name.

NOTE: The nested directories are unchanged and retain the met field in
the directory name. This is because the GEOS-5, GEOS-FP, and MERRA-2
nested grids have different horizontal resolutions and so the run
directories are not interchangeable.

In this commit, the following changes are also made:

  - The benchmark simulation is now referred to as the "standard" instead
    of "benchmark." This should make it more clear to users that this is
    the recommended GEOS-Chem simulation. The files in chem_inputs/benchmark
    have been moved to chem_inputs/standard.

  - Symbolic links for HEMCO_restart have been removed. This forces the
    user to use the default values specified by HEMCO and to spin up a
    new HEMCO restart file. The HEMCO_restart file originally stored in
    geos5_4x5_tropchem is now stored in 4x5_standard, so that the 1-month
    benchmark simulations will be unaffected.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>