Add fixes for unit text matrix web page
The old code relied on the listing of $runRoot to set up the unitTests
hash that contains the results of the unit test for displaying in txt and
html files. This method can no longer be used because the run directories
have been consolidated so that we have one directory (e.g. 4x5_tropchem)
instead of four (e.g. geos5_4x5_tropchem, merra_4x5_tropchem). This also
messes up replacing the utName string with the utColor.

To resolve these issues, ut_template.html has been updated so that the
cells of the unit test matrix in the html file are now automatically set
to white (bgcolor="#FFFFFF"). Routine makeMatrix now checks to see if
utName (obtained from the results.log file) matches the name of the
simulation specified in the html file. If they match, the background color
will be replaced with utColor.

Also updated ut_template.html so that standard, aciduptake, and marinePOA
simulations are listed in the main table instead of the table labeled
'Special UnitTests' to reflect that these simulations can now be easily
run with GEOS-5, GEOS-FP, MERRA, or MERRA-2 met fields.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>