Remove blank placeholder text files in UT/runs/
Several blank text files were located in the UT/runs directory as
placeholders for simulations that didn't have run directories. These were
files were likely added so that the unitTests perl hash could be populated
with a complete list of possible unit tests, however we no longer use the
run directory names to do that. There doesn't seem to be a need for these
files anymore, so they have been removed to avoid user confusion. The
exceptions are the 4x5_CO2 and 2x25_TOMAS40 text files, which inform the
users that those run directories don't exist in the unit tester but they
can be created from the 2x25_CO2 and 4x5_TOMAS40 directories, respectively.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>