Make sure temporary input.geos files are removed when executing cleanFiles
Added input.geos.g* and input.geos.m* to the list of files that are
removed when executing perl/cleanFiles and perl/cleanRunDirs. This ensures
that any temporary input.geos files created from the input.geos.template
files are removed. The g* and m* ensures removal of all combinations of
met_grid_sim.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>