Turn prod/loss diagnostic back on for the TOMAS simulations
The ND65 prod/loss diagnostic has been restored in the FlexChem code.
We can now obtain this information from KPP and output it to the bpch
file.

NOTE: The TOMAS simulation will still crash and fail unit tests because
of outstanding issues in the FlexChem implementation. We hope to resolve
this prior to the v11-01 release.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>