Add fix to perl/gcCreateDiffTest for replacing text in bash scripts
In the bash scripts summarizeDiagDiffs.sh and locateDiagDiffs.sh, the
MET_GRID_SIM string needs to be replaced with $met_$grid_$sim and not
the original run directory name (i.e. $grid_sim). Without this fix, the
two bash scripts would crash because they couldn't find the files to
compare.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>