Update gcCreateDiffTest to replace N_TRACERS string in PlotDiff.input
Most simulations have a unique number of tracers. To accommodate creation
of difference plots for the common simulations that the GCST does difference
tests for, we now hardcode nTracers in gcCreateDiffTest and replace the
string N_TRACERS in PlotDiff.input with that value. In the future, we can
try to get the number of tracers in a more elegant way (from input.geos).

Also, update plot_diffs.pro to now pass keyword NO_FULLCHEM to routine
benchmark_1mon. This will allow for for creation of difference plots for
specialty simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>