Add difftest/Ref/resetLinks.pl script
The resetLinks.pl script will be copied into the Ref folder
of difference-test directories generated by the script
perl/gcCreateDiffTest.

When you manually copy a difference-test directory to a new
location, the links in the Ref directory still point to the
old location. Running Ref/resetLinks.sh will remove the old
links and create new links to the input files in the ../Dev
folder.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>