Use ifort13/OpenMPI/odyssey as default settings in GCHP run directory
Performance is improved with ifort15 on Odyssey, or by using MPAVICH2.
However, there are limited licenses available for that compiler on Odyssey
and the default source code is set for OpenMPI. Updating the default
MPI implementation in the run directory should occur concurrently with
updating the default MPI implementation hard-coded in the GCHP source
code.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>