Add workaround for Git versions prior to v1.8.5
Git v1.8.5 and higher use the "git -C" option to point to code
directories other than the current directory.  This -C option is used
in the run-directory Makefile commands used to obtain information
about the last committed state of the code.

For older Git versions, we have implemented a workaround where we
manually change to the code directory, issue a command to check the
last commited state, and then change back.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>