Add workaround for older Git versions to the diffteset/Makefile
We have now added the same workaround for Git versions older than
v1.8.5 (added to the the prior commit 827fe62) to the difference
test makefile: perl/difftest/Makefile.

This will make sure that the difftest build sequence will be able to
get info about the last committed state of the Def and Ref codes
for all versions of Git.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>