Update the 4x5_standard/v11-01.run script
Now use #SLURM -c 24 and $SLURM_CPUS_PER_TASK, which is recommended
by FASRC staff for OpenMP simulations.

Also use srun -c $OMP_NUM_THREADS in the line that calls GEOS-Chem,
which will tell SLURM to restrict the multi-threading to that
line only (also per FASRC staff).

Change time to time -p, which will output the result in seconds,
which is much easier to use in scripts that examine the time output.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>