Initial commit of gchp_4x5_standard to work with v11-01; not yet functional
Files in the gchp_4x5_standard directory are ready for use with GCHP_Dev
except for HEMCO_Config.template and ExtData.rc (and permutations). Both
of these files need updating to combine Mike's updates with the latest
settings in GC classic rundir 4x5_standard/. Copies of Mike's files are
included in this commit for reference.

Also in this commit: remove outdated files in gchp_4x5_tropchem that
were brought back in with the Dev merge.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>