Switched GCHP resolution and chemistry definitions
Previously, two GCHP run directories (standard and tropchem) were
available. However, it seemed easier to maintain only a "master"
mechanism (standard) during initial development, and then
update the "tropchem" mechanism at a later date. As such, only
the standard mechanism remains. The GCHP resolution string has
been changed from "4x5" to the equivalent "c24", and the GCHP
run directory has been updated and substantially simplified.
The primary outcome of this is that the new default directory
sets - and checks for - ESMF_COMM and MPI_ROOT, which is all that
is now required to set the MPI distribution when compiling GCHP.
The choice of meteorological data is also restricted to 2x2.5, on
the basis that users can easily modify this should they so wish.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>