Update run directory Makefile to allow for compiling with TAU_PROF=y
The file perl/Makefile is now updated to pass TAU_PROF to the list of
make options. When TAU_PROF=y, TAU_OPT will be set according to the
TAU_OPTIONS environment variable.

The run directory Makefile is also updated to set TRACEBACK=y by default.
This is what is done in the GEOS-Chem source code, so we do it here for
consistency and so the correct information is printed to the lastbuild
file.

Also update perl/difftest/Makefile to be consistent with perl/Makefile.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>