Add complete list of run directory options to CopyRunDirs.input
Some users may have been confused that CopyRunDirs.input was meant only as
an example file and not all possible run directories were listed there. To
avoid confusion, we now list most, if not all, of the run directories that
can be made with the GEOS-Chem unit tester. This should be consistent with
the entries in the default UnitTest.input file. The end dates in the
CopyRunDirs.input file have also been updated so by default the run will
be set up for a one-month simulation.

Several UnitTest.input files have also been updated here to (1) activate
the GEOS-5 0.5x0.666 nested NA mercury simulation, (2) to rename the
"GEOS-Chem benchmark" header to "GEOS-Chem standard simulation," and (3)
move the GEOS-Chem standard simulations to the top of the list to make it
clear that the standard simulation is the recommended GEOS-Chem simulation.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>