Update perl scripts to create symbolic links to GEOS-Chem restart files
In the previous commit, the file path to the SPC_RESTARTS directory was
removed so that GEOS-Chem instead looks for GEOSChem_restart.YYYYMMDDhhmm.nc
in the run directory. Here, gcUnitTest, gcCopyRunDirs, and gcCreateDiffTest
are updated to create a symbolic link called GEOSChem_restart.${start}.nc
that will point to the appropriate sample restart file in SPC_RESTARTS. If
no sample restart file is found, a warning will print to the screen for
gcCopyRunDirs and gcCreateDiffTest informing the user that no restart file
will be included in their run directory. The gcUnitTest script will stop
with an error if no sample restart file is found.

Also made the following minor updates in this commit:

(1) Add masscons entries to the CopyRunDirs.input file.
(2) Add complete list of run directory options to DiffTest.input.
(3) Rename variables in perl/Makefile for clarity. TRAC_AVG is now GC_DIAG
    and TRAC_RST is now GC_RST following the naming convention that is/will
    be used for the netCDF output.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>