Add passive species menu to input.geos files for RnPbBe simulations
The PASV species was to the Rn-Pb-Be simulation early in v11-01 to validate
transport and mass conservation. Since then Christoph has added a passive
species module that easily allows for the addition of multiple passive
species. To avoid redundant code, PASV has been removed from the species
database and will instead be declared in the passive species menu of
input.geos. The passive species PASV is setup with MW = 1 g/mol, unlimited
atmospheric lifetime, and an initial concentration of 100 ppb globally,
following how PASV was set up in the GEOS-Chem code.

Also turn on ND49 by default so that the mass of the passive species is
saved out to species_mass_kg.txt every 6 hours.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>