Add minor fixes for creating symbolic links to GEOS-Chem restart files
Force creation of a symbolic link pointing to GEOS-Chem restart files in
SPC_RESTARTS even if a symbolic link already exists. Not including -f in
the ln command caused the Unit Tester to crash with an error if cleanFiles
wasn't executed prior to submitting a new unit test and old symbolic links
remained.

Also added GEOSChem_restarts* to the list of files removed when executing
cleanRunDirs, following what is done in cleanFiles.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>