Updated GCHP HEMCO_Config file to match GCC
The GCHP HEMCO_Config template was no longer up to date. It has now
been updated to match the GCC 4x5 standard template, with a few
modifications necessary for the different resolution and for some
of the additional files used to drive GCHP (e.g. the GCHP MODIS LAI
files).

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>