Strip out " characters from the LAST_COMMIT string in Makefiles
We recently discovered an unexpected bug.  If a commit string has
a " character in it, then this will cause an error when trying to
print out the Makefile varaible LAST_COMMIT..  GNU Make will
interpret the " character as terminating the LAST_COMMIT string
prematurely, which throws an error.

The fix is just to strip out any such " characters from LAST_COMMIT,
to avoid such errors in the future.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>