Update OTD-LIS scaling and dust mass tuning factor to best match GCC
Seb Eastham submitted the updated values. He writes: "This corresponds to
an annual average total flash rate of 145.5 flashes/second in GCHP, and
a dust mass flux rate in GCHP at C24 which is 8.34% greater than that
observed in GCC at 4x5. This is assuming that the annual average ratio
is roughly accurate for the target time period."

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>