Add a HEMCO standalone run directory (4x5_HEMCO)
Christoph Keller provided a sample HEMCO standalone run directory for
inclusion in the GEOS-Chem Unit Tester. Users can create a HEMCO standalone
run directory by adding or uncommenting the line:

  geosfp   4x5         -      HEMCO     2013070100   2013080100   -

in their CopyRunDirs.input file. This will copy the run directory to the
specified path and also create HEMCO_Config.rc and HEMCO_sa_Config.rc files
from the template files.

NOTE: The Makefile that comes with the HEMCO run directory is meant to be
used to build and copy the HEMCO standalone executable (hemco_standalone.x)
to the run directory. However, some modification is still needed in the
GEOS-Chem Makefiles to allow for compilation of HEMCO source code only.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>