Modify perl/getRunInfo for the HEMCO standalone simulations
Now return the start and end dates (options 4 and 5) in YYYYMMDDhhmm
format for the HEMCO standalone.

Prior to this commit, the start and end dates would be returned in
YYYY-MM-DD hh:mm:ss format.  This is problematic because the Makefiles
expect a string in YYYYMMDDhhmm format for use in the HEMCO restart
file names.  Now fixed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>